+++ example/container.py
@@ -2,7 +2,7 @@ import cadquery as cq
 from cqportal import Container
 
 bp_container = Container()
-bp_container.bp_hinge.rotate_deg = 0
+bp_container.bp_hinge.rotate_deg = -90
 
 bp_container.make()